refactor(ipc): split dictation tests into tests.rs (#684)
Move the `#[cfg(test)] mod tests { ... }` block out of
`commands/dictation/mod.rs` into a sibling `tests.rs` file.

No behaviour change — all 16 dictation tests continue to pass
under `ipc::commands::dictation::tests::`. The inline test
module was the only thing keeping `dictation/mod.rs` at 1155 lines;
the command shells now sit in a focused 450-line file.

Import strategy in tests.rs: explicit `use super::pipeline::{...}`
imports replace the implicit pull-through that `use super::*` gave
when the tests lived inside `mod.rs`. `ForegroundApp` is imported
directly from `crate::ipc::state` so the `#[cfg(test)] use
super::ForegroundApp` shim in `mod.rs` can be removed.
